+++ js/src/admin/index.js
@@ -117,6 +117,17 @@ app.initializers.add('pianotell-flamoji', (app) => {
           </div>
           <div className="Flamoji--emojiSetting">
             <div className="Form-group">
+              <Switch
+                state={!!this.setting(['pianotell-flamoji.sticker_mode'])() && this.setting(['pianotell-flamoji.sticker_mode'])() !== '0'}
+                onchange={this.settings['pianotell-flamoji.sticker_mode']}
+              >
+                {app.translator.trans('pianotell-flamoji.admin.settings.sticker_mode_label')}
+              </Switch>
+            </div>
+            <div className="helpText">{app.translator.trans('pianotell-flamoji.admin.settings.sticker_mode_text')}</div>
+          </div>
+          <div className="Flamoji--emojiSetting">
+            <div className="Form-group">
               <label>{app.translator.trans('pianotell-flamoji.admin.settings.picker_set_label')}</label>
               <Select
                 value={this.setting(['pianotell-flamoji.picker_set'])() || 'auto'}